+++ src/map/packets/change_music.h
@@ -38,7 +38,7 @@ class CChangeMusicPacket : public CBasicPacket
 {
 public:
 
-    CChangeMusicPacket(uint8 BlockID, uint32 MusicTrackID);
+    CChangeMusicPacket(uint8 BlockID, uint8 MusicTrackID);
 };
 
 #endif
